refactor: move the duplicate functions into utils
